Move Sapling transaction builder to an independent module.
